+++ nextapp/src/util/GoogleAnalytics.tsx
@@ -1,17 +1,17 @@
 'use client';
-
+import { NextScript } from 'next/document';
 import Script from 'next/script'
 
 export default function GoogleAnalytics({GA_TRACKING_ID} : {GA_TRACKING_ID : string}){
     return (
         <>
 					<Script 
-						strategy="lazyOnload" 
+						strategy='lazyOnload'
 						src={`https://www.googletagmanager.com/gtag/js?id=${GA_TRACKING_ID}`}
 					/>
 					<Script 
 						id='google-analytics' 
-						strategy="lazyOnload"
+						strategy='lazyOnload'
 						dangerouslySetInnerHTML={{
 							__html: `
 							window.dataLayer = window.dataLayer || [];
